+++ src/classes/UI5Classes/JSParser/strategies/FieldsAndMethodForPositionBeforeCurrentStrategy.ts
@@ -13,9 +13,10 @@ export class FieldsAndMethodForPositionBeforeCurrentStrategy extends FieldMethod
 		const UIClassName = this.getClassNameOfTheVariableAtPosition(className, offset);
 		if (UIClassName) {
 			fieldsAndMethods = this.destructueFieldsAndMethodsAccordingToMapParams(UIClassName);
-			const classNameOfTheCurrentDocument = AcornSyntaxAnalyzer.getClassNameOfTheCurrentDocument();
-			if (fieldsAndMethods && classNameOfTheCurrentDocument !== className) {
+			if (fieldsAndMethods && className !== fieldsAndMethods.className) {
 				this._filterFieldsAndMethodsAccordingToAccessLevelModifiers(fieldsAndMethods);
+			} else if (fieldsAndMethods) {
+				this._filterFieldsAndMethodsAccordingToAccessLevelModifiers(fieldsAndMethods, ["private", "protected", "public"])
 			}
 		}
 
@@ -36,7 +37,8 @@ export class FieldsAndMethodForPositionBeforeCurrentStrategy extends FieldMethod
 					name: field,
 					description: field,
 					type: "any",
-					visibility: "public"
+					visibility: "public",
+					owner: ""
 				}))
 			};
 		} else if (classNamePartsFromMapParam.length > 1) {
@@ -54,7 +56,8 @@ export class FieldsAndMethodForPositionBeforeCurrentStrategy extends FieldMethod
 							description: key,
 							name: key,
 							visibility: "public",
-							type: typeof paramStructure[key] === "string" ? paramStructure[key] : typeof paramStructure[key]
+							type: typeof paramStructure[key] === "string" ? paramStructure[key] : typeof paramStructure[key],
+							owner: ""
 						};
 					}) : [],
 					methods: []