+++ src/classes/providers/diagnostics/js/jslinter/parts/config/ConfigHandler.ts
@@ -5,9 +5,10 @@ export class ConfigHandler {
 		return vscode.workspace.getConfiguration("ui5.plugin").get("JSLinterExceptions") || [];
 	}
 
-	static checkIfMethodNameIsException(className = "", memberName = "") {
+	static checkIfMemberIsException(className = "", memberName = "") {
+		const hardcodedExceptions = ["metadata", "renderer", "onAfterRendering", "customMetadata"];
 		const classExceptions = ConfigHandler.getJSLinterExceptions();
-		const isException = !!classExceptions.find(classException => {
+		const isException = hardcodedExceptions.includes(memberName) || !!classExceptions.find(classException => {
 			let isException = (classException.className === className || classException.className === "*") &&
 				(classException.memberName === memberName || classException.memberName === "*");