+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.12.26 (22-04-2021)
+* Bugfixes
+* Public member linter introduced
+* Unused method linter now is replaced with unused member linter, which also checks fields
+
 ## 0.12.25 (21-04-2021)
 * Bugfixes